+++ app/common/routes.jsx
@@ -4,13 +4,13 @@ import { Route, IndexRoute } from 'react-router'
 import AppLayout from './layouts/app'
 import requireAuthentication from './../auth/components/protected_route'
 
-import HelloWorld from './../hello_world/views/hello'
+import DecksOverview from './../srs/views/decks_overview'
 import AuthLogin from './../auth/views/login'
 import AuthRegister from './../auth/views/register'
 
 export default (
   <Route path="/" component={AppLayout}>
-    <IndexRoute component={HelloWorld}/>
+    <IndexRoute component={DecksOverview}/>
     {'/* Auth */'}
     <Route path="login" component={AuthLogin}/>
     <Route path="register" component={AuthRegister}/>
