+++ app/srs/components/DeckItem.jsx
