+++ app/srs/reducer.js
