+++ app/srs/styles/decks_overview.scss
