+++ app/srs/views/decks_overview.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import DecksOverviewContainer from '../containers/DecksOverview'
+
+function DecksOverview() {
+  return (
+    <DecksOverviewContainer/>
+  )
+}
+
+
+export default DecksOverview
